Not enable the upcoming feature: ExistencialAny
Tuist builds the project for release using Xcode 14 to ensure ProjectDescription
works with that Xcode version upward. When we use Xcode 15 for development,
the Swift version is >= 5.9 and that causes this package to have a setting
that's not supported when building for Xcode 14. As a result, the release builds fail.